add reloadCommands, modify deleteCommand

+++ src/slash_commands/admin/deleteCommand.js
@@ -1,27 +1,23 @@
 const { client } = require("../../index");
 
 const execute = async (message, args) => {
-  client.api.applications(client.user.id).guilds(process.env.GUILD_ID).commands.get().then(commands => {
-    commands.forEach(command => {
-      if (command.name === args[0]) {
-        client.api.applications(client.user.id).guilds(process.env.GUILD_ID).commands(command.id).delete();
-      }
+  if (message.member.hasPermission("ADMINISTRATOR")) {
+    client.api.applications(client.user.id).guilds(process.env.GUILD_ID).commands.get().then(commands => {
+      commands.forEach(command => {
+        if (command.name === args[0]) {
+          client.api.applications(client.user.id).guilds(process.env.GUILD_ID).commands(command.id).delete();
+        }
+      });
     });
-  });
+  }
 };
 
 module.exports = {
   prefix: true,
   name: "deletecommand",
-  description: "Delete a slash command, mostly for development purposes",
+  description: "Delete a slash command",
   role: "admin",
-  options: [
-    {
-      name: "command",
-      description: "Command to delete",
-      type: 3,
-      required: true,
-    },
-  ],
+  usage: "[command name]",
+  args: true,
   execute,
 };
